+++ backend/src/api/endpoints/api_playlist_utils.rs
@@ -188,6 +188,7 @@ mod tests {
             category_id: 0,
             rating: 0.0,
             input_name: "test".intern(),
+            epg_channel_id: None,
         }
     }